==> notebooks/1. First Clean.ipynb <==
from pyspark.sql import SparkSession
from urllib.request import urlretrieve
# Create a spark session (which will run spark jobs)
spark = (
    SparkSession.builder.appName("ADS project 1")
    .config("spark.sql.repl.eagerEval.enabled", True) 
    .config("spark.executor.memory","2G")
    .config("spark.driver.memory","4G")
    .config("spark.sql.parquet.cacheMetadata", "true")
    .config("spark.sql.session.timeZone", "Etc/UTC")
    .getOrCreate()
)

spark.conf.set("spark.sql.parquet.compression.codec", "gzip")

from pyspark.sql import functions as F
from pyspark.sql import SparkSession
from pyspark.sql.functions import *
import matplotlib.pyplot as plt
import pandas as pd
import numpy as np
import seaborn as sbs
import geopandas as gpd
import folium

# The timeline of year and month of data to use
YEARS = ['2019', '2021']
MONTHS = range(1, 13)
path_raw = f"../data/raw/tlc_data/"
path_curated = f"../data/curated/tlc_data/"

# read in the data
sdf_2019 = spark.read.parquet(f'{path_raw}2019/')
sdf_2021 = spark.read.parquet(f'{path_raw}2021/')

# brief overview of the final response feature
sdf_2021.select('tip_amount').describe()

# add the features that we need for the analysis based on the some of original data attributes
def add_feature(sdf):
    return sdf.withColumn('Month,Date', date_format("tpep_pickup_datetime", "MM,dd"))\
        .withColumn('Month', date_format("tpep_pickup_datetime", "MM").cast('int'))\
        .withColumn('Pickup_time', date_format('tpep_pickup_datetime', 'HH').cast('int'))\
        .withColumn("time_duration", unix_timestamp('tpep_dropoff_datetime') - unix_timestamp('tpep_pickup_datetime'))\
        .withColumn("is_weekend", dayofweek("tpep_pickup_datetime").isin([1,7]).cast("int"))\
        .withColumn("average_speed", F.col('trip_distance') / (F.col('time_duration')/3600))\
        .withColumn('is_airport', (F.col('DOLocationID')==1) | (F.col('DOLocationID')==132) | (F.col('DOLocationID')==138)\
                | (F.col('PULocationID')==1) | (F.col('PULocationID')==132) | (F.col('PULocationID')==138))

# remove the data records that are invalid against common sense or business rule
def clean_feature(sdf):
    return sdf.where(F.col('time_duration') >= 60).where(F.col('tip_amount') > 0)\
            .where(F.col('trip_distance') > 0).where(F.col('payment_type') == 1)\
            .where(F.col('fare_amount') >= 2.5)

# for data of each month in each year, add features and do the first cleaning then save them into the folder of '/firstclean'
for year in YEARS:
    for month in MONTHS:
        # turn '1' to '01' to keep consistent with the format of the file names
        month = str(month).zfill(2)
        sdf = spark.read.parquet(f'{path_raw}{year}/{year}-{month}.parquet')
        # add the features
        sdf = add_feature(sdf)
        # filter the invalid
        sdf = clean_feature(sdf)
        # write the cleaned data to directory
        sdf.write.mode("overwrite").parquet(f'{path_curated}{year}/firstclean/{year}--{month}-firstclean.parquet')
        print(f"finished {year} {month}")
