==> notebooks/5. external_data.ipynb <==
import pandas as pd
import numpy as np

YEARS = ['2019', '2021']

# format the external dataset
def process_external(weather, year):
    for i in weather.columns:
        if i != 'Month,Date':
            # change the data types into numeric
            weather[i] = pd.to_numeric(weather[i], errors='raise', downcast=None)
    # decide whether it is raining based on precipitation
    weather['is_rainy'] = weather['Precipitation (in)'] != 0
    weather.to_csv(f'../data/curated/NYC weather {year} cleaned.csv')
    print(f'finished {year}')

for year in YEARS:
    # preprocess the two external datasets
    weather = pd.read_csv(f"../data/raw/NYC weather {year}.csv")
    process_external(weather, year)
