==> notebooks/11. Model Data.ipynb <==
from pyspark.sql import SparkSession
from urllib.request import urlretrieve
# Create a spark session (which will run spark jobs)
spark = (
    SparkSession.builder.appName("ADS project 1")
    .config("spark.sql.repl.eagerEval.enabled", True)  
    .config("spark.executor.memory","4G")
    .config("spark.driver.memory","8G")
    .config("spark.sql.parquet.cacheMetadata", "true")
    .config("spark.sql.session.timeZone", "Etc/UTC")
    .getOrCreate()
)

spark.conf.set("spark.sql.parquet.compression.codec", "gzip")

from pyspark.sql import functions as F
from pyspark.sql import SparkSession
from pyspark.sql.functions import *
import matplotlib.pyplot as plt
import pandas as pd
import numpy as np
import seaborn as sbs
import geopandas as gpd
import folium
from pyspark.ml.feature import OneHotEncoder
from pyspark.ml.feature import VectorAssembler
from pyspark.ml.feature import MinMaxScaler
from pyspark.ml import Pipeline
from pyspark.sql.functions import udf
from pyspark.sql.types import DoubleType
from pyspark.ml.regression import RandomForestRegressor
from pyspark.ml.evaluation import RegressionEvaluator

# files that are going to be used in modelling
files = ['../data/curated/tlc_data/2021/final_data/2021--06-final_merged.parquet', \
        '../data/curated/tlc_data/2021/final_data/2021--07-final_merged.parquet', \
        '../data/curated/tlc_data/2021/final_data/2021--08-final_merged.parquet',\
        '../data/curated/tlc_data/2021/final_data/2021--09-final_merged.parquet',\
        '../data/curated/tlc_data/2021/final_data/2021--10-final_merged.parquet',\
        '../data/curated/tlc_data/2021/final_data/2021--11-final_merged.parquet']

# file that is going to be predicted
predict = '../data/curated/tlc_data/2021/final_data/2021--12-final_merged.parquet'

# final change of datatype to meet the requirements of the modelling
def change_type(sdf):
        return sdf.withColumn('is_airport', col('is_airport').cast("integer"))\
                .withColumn('is_rainy', (col('is_rainy') == 'True').cast('integer'))\
                .withColumn('Temperature (F)', col('Temperature (F)').cast('float'))\
                .withColumn('Wind Speed (mph)', col('Wind Speed (mph)').cast('float'))

# encode the categorical features so that they could fit into a numeric model
def one_hot(sdf):
    ohe = OneHotEncoder()
    # attributes that need to be encoded
    ohe.setInputCols(["is_weekend", 'is_airport', 'is_rainy', 'Pickup_time', 'Month'])
    ohe.setOutputCols(["Is_Weekend", "Is_Airport", "Is_Rainy", 'Pickup_Time', 'MONTH'])
    model = ohe.fit(sdf)
    # encode the dataset
    sdf = model.transform(sdf)
    return sdf

path = "../data/curated/tlc_data/"
# read in all files for training
sdf = spark.read.parquet(*files)
sdf = change_type(sdf)
sdf = one_hot(sdf)
# save as the final model data
sdf.write.mode("overwrite").parquet(f'{path}/Model_data.parquet')

# read in the file for testing
sdf = spark.read.parquet(predict)
sdf = change_type(sdf)
sdf = one_hot(sdf)
# save sa the final test data
sdf.write.mode("overwrite").parquet(f'{path}/Test_data.parquet')
