==> notebooks/13. Classification.ipynb <==
from pyspark.sql import SparkSession
from urllib.request import urlretrieve
# Create a spark session (which will run spark jobs)
spark = (
    SparkSession.builder.appName("ADS project 1")
    .config("spark.sql.repl.eagerEval.enabled", True)
    .config("spark.executor.memory","4G")
    .config("spark.driver.memory","8G")
    .config("spark.sql.parquet.cacheMetadata", "true")
    .config("spark.sql.session.timeZone", "Etc/UTC")
    .getOrCreate()
)

spark.conf.set("spark.sql.parquet.compression.codec", "gzip")

from pyspark.sql import functions as F
from pyspark.sql import SparkSession
from pyspark.sql.functions import *
import matplotlib.pyplot as plt
import pandas as pd
import numpy as np
import seaborn as sbs
import geopandas as gpd
import folium
from pyspark.ml.feature import OneHotEncoder
from pyspark.ml.feature import VectorAssembler
from pyspark.ml.feature import MinMaxScaler
from pyspark.ml import Pipeline
from pyspark.sql.functions import udf
from pyspark.sql.types import DoubleType
from pyspark.ml.regression import RandomForestRegressor
from pyspark.ml.evaluation import RegressionEvaluator
from pyspark.ml.regression import GeneralizedLinearRegression
from pyspark.ml.regression import LinearRegression

FEATURES = ['PULocationID', 'trip_distance', 'fare_amount', 'average_speed', 'time_duration', \
   'Temperature (F)', 'Wind Speed (mph)', 'Pickup_Time', 'Is_Airport', 'Is_Weekend', 'Is_Rainy']
features = 'features'

# vectorise the features that are going to be put into the model
def feature_converter(sdf):
    vecAss = VectorAssembler(
    # features to be used
    inputCols=FEATURES, 
    # name of the output column
    outputCol=features
    )
    df_va = vecAss.transform(sdf)
    return df_va

# act as a pipline before modeling
def preparation(sdf):
    # vectorisation
    sdf = feature_converter(sdf)
    # take only the feature vector and the response
    sdf = sdf.select([features, 'tip_amount'])
    return sdf

# read in the train test dataframe
train_sdf = spark.read.parquet('../data/curated/tlc_data/Model_data.parquet')
test_sdf = spark.read.parquet('../data/curated/tlc_data/Test_data.parquet')
# pipline the dataframes
train_sdf = preparation(train_sdf)
test_sdf = preparation(test_sdf)

# discretisation the Tip amount into labels of High and Low
from pyspark.sql.types import IntegerType
@F.udf(IntegerType())
def discretisation(tip):
    if tip <= 2:
        return 0
    else:
        return 1


# apply the discretisation
train_sdf = train_sdf.withColumn('label', discretisation(F.col('tip_amount')))
test_sdf = test_sdf.withColumn('label', discretisation(F.col('tip_amount')))

# baseline accuracy
train_sdf.where(F.col('label') == 1).count()/train_sdf.count()

from pyspark.ml.classification import LogisticRegression

#Apply the logistic regression model

log_reg=LogisticRegression(labelCol='label', featuresCol = features, elasticNetParam=1, family="binomial").fit(train_sdf)
result = log_reg.transform(test_sdf)

# define the values for error analysis
TN = result.where((F.col('label') == 0) & (F.col('prediction') == 0)).count()
FP = result.where((F.col('label') == 0) & (F.col('prediction') == 1)).count()
FN = result.where((F.col('label') == 1) & (F.col('prediction') == 0)).count()
TP = result.where((F.col('label') == 1) & (F.col('prediction') == 1)).count()

# produce evaluation scores
acc = (TN + TP)/(TN + TP + FN + FP)
recall = TP/(TP + FN)
precision = TP / (TP + FP)
F1 = 2*recall*precision / (recall + precision)
print(f'The Accuracy is: {acc}')
print(f'The Recall is: {recall}')
print(f'The Precision is: {precision}')
print(f'The F1 score is: {F1}')
